refactor: 서버 선택 ui 컴포넌트 분리
